move test to shared adapter zone for reuse
